+++ packages/commerce-sdk-react/src/hooks/ShopperCustomers/index.test.ts
@@ -33,7 +33,6 @@ describe('Shopper Customers hooks', () => {
         // If this test fails: add cache update logic, remove the endpoint from the mutations enum,
         // or add it to the `expected` array to indicate that it is still a TODO.
         expect([...unimplemented]).toEqual([
-            'updateCustomerPassword',
             'deleteCustomerProductList',
             'updateCustomerProductList'
         ])